poster/pyproject.toml
[tool.poetry]
name = "poster"
version = "0.1.0"
description = ""
authors = ["Frank Wickstroem <frank.wickstrom@anders.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.61.1"
python-dotenv = "^0.15.0"
uvicorn = {version = "^0.12.2", extras = ["standard"]}
gunicorn = "^20.0.4"
aio-pika = "^6.7.1"
slackclient = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
